requirements.txt
openai>=1.0.0
google-auth>=2.0.0
requests>=2.25.0
fastapi
uvicorn
httpx
pydantic
